repo/backend/requirements.txt
fastapi[standard]>=0.113.0,<1.0.0
gunicorn>=22.0.0,<23.0.0
pydantic-settings>=2.2.1,<3.0.0
jinja2>=3.1.0,<4.0.0
pymysql>=1.1.1,<2.0.0
bcrypt>=4.1.3,<5.0.0
